All/feature/unst 9468 fortran styler array delimiters (#414)
* Rename the double_precision_conversion folder to deltares_fortran_styler

* Rename automatic_conversions to fortran_styler

* Refactor double precision conversion script into larger fortran styler script, add array delimiter conversion

* Convert dflowfm_kernel code again, new converter found some more double precision cases

* Add unit test structure for fortran styler, add unit tests for double precision conversion, fix bug with newlines in added use statements

* Add unit tests for array delimiter converter, fix bug with trailing whitespace inside array literal

* Fix tool to not use python 3.7 features, since alma 8 requires python 3.6

* Add build step to run the unit tests for the fortran styler

* Update project structure to conventional Python project layout

* Fix paths after project layout restructuring, ensure that it can still run as a script. Fix bug in double precision converter

+++ src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_utils/solve_parms.F90
@@ -43,12 +43,12 @@ module m_parms
    parms_PC :: pc ! preconditioner
    parms_Solver :: ksp ! solver
 
-   double precision, dimension(:), allocatable :: sol ! solution vector
-   double precision, dimension(:), allocatable :: rhs ! right-hand side vector
+   real(kind=dp), dimension(:), allocatable :: sol ! solution vector
+   real(kind=dp), dimension(:), allocatable :: rhs ! right-hand side vector
 
    integer :: numloc ! number of local rows (elements) in matrix
    integer :: nummat ! number of matrix entries
-   double precision, dimension(:), allocatable :: amat ! matrix entries, dim(nummat)
+   real(kind=dp), dimension(:), allocatable :: amat ! matrix entries, dim(nummat)
    integer, dimension(:), allocatable :: imat, jmat ! column indices and row pointers of matrix, dim(numloc+1) and dim(nummat) respectively
    integer, dimension(:), allocatable :: guusidx ! index in guus matrix-array
 
@@ -306,10 +306,10 @@ subroutine conjugategradient_parms(s1, Ndx, its)
    implicit none
 
    integer, intent(in) :: Ndx !< number of unknowns (flownodes)
-   double precision, dimension(Ndx), intent(inout) :: s1 !< water level
+   real(kind=dp), dimension(Ndx), intent(inout) :: s1 !< water level
    integer, intent(out) :: its !< number of iterations
 
-   double precision :: res ! residual norm
+   real(kind=dp) :: res ! residual norm
 
    integer :: i, j, n
